pyproject.toml
[project]
name = "string-lights"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "torch>=2.0",
    "transformers>=4.30",
    "Pillow>=10.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
string-lights = "string_lights.__main__:main"
